Delete pet with authorization

+++ src/actions/actions.ts
@@ -103,6 +103,14 @@ export async function editPet(petId: unknown, newPetData: unknown) {
 export async function deletePet(petId: unknown) {
   await sleep(1);
 
+  // authentication check
+  const session = await auth();
+
+  if (!session?.user) {
+    redirect("/login");
+  }
+
+  // validation check
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -110,6 +118,27 @@ export async function deletePet(petId: unknown) {
       message: "Invalid pet data",
     };
   }
+
+  //authorization check
+  const pet = await prisma.pet.findUnique({
+    where: {
+      id: validatedPetId.data,
+    },
+  });
+
+  if (!pet) {
+    return {
+      message: "Pet not found",
+    };
+  }
+
+  if (pet.userId !== session.user.id) {
+    return {
+      message: "Unauthorized",
+    };
+  }
+
+  //database mutation
 
   try {
     await prisma.pet.delete({
